requirements.txt
aiohttp==3.7.4
requests==2.24.0
pika==1.1.0
flake8==3.8.4
autopep8==1.5.4
flake8-import-order==0.18.1
python-dotenv==0.19.0
accelerate==0.26.1
certifi==2023.11.17
charset-normalizer==3.3.2
diffusers==0.25.1
filelock==3.13.1
flask==3.0.1
fsspec==2023.12.2
huggingface-hub==0.20.3
idna==2.5
importlib-metadata==7.0.1
Jinja2==3.1.3
MarkupSafe==2.1.4
minio==7.2.3
mpmath==1.3.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pillow==10.2.0
psutil==5.9.8
PyYAML==6.0.1
regex==2023.12.25
safetensors==0.4.2
sympy==1.12
tokenizers==0.15.1
torch==2.1.2
tqdm==4.66.1
transformers==4.37.0
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.24.1
zipp==3.17.0
